Use qunit-node for running tests
