--- test/test-runner.html
@@ -1,21 +1,0 @@
-<!doctype html>
-
-<html lang="en">
-<head>
-  <meta charset="utf-8">
-  <title>toMarkdown test runner</title>
-  <link rel="stylesheet" href="lib/qunit.css"></link>
-  <script src="https://ajax.googleapis.com/ajax/libs/jquery/1.6.4/jquery.min.js"></script>
-  <script src="lib/qunit.js"></script>
-  <script src="../src/to-markdown.js"></script>
-  <script src="tests.js"></script>
-</head>
-<body>
-  <h1 id="qunit-header">toMarkdown Tests</h1>
-  <h2 id="qunit-banner"></h2>
-  <div id="qunit-testrunner-toolbar"></div>
-  <h2 id="qunit-userAgent"></h2>
-  <ol id="qunit-tests"></ol>
-  <div id="qunit-fixture">test markup, will be hidden</div>
-</body>
-</html>
